Restructured template trees and tests
