+++ fab/ThrowableDiagnosticV1Decorators/NestedDiagnosticV1/NestedDiagnosticDecorator.service.yml
@@ -3,5 +3,3 @@ services:
     class: Neighborhoods\ThrowableDiagnosticComponent\ThrowableDiagnosticV1Decorators\NestedDiagnosticV1\NestedDiagnosticDecorator
     public: false
     shared: false
-    calls:
-      - [setDiagnosedFactory, ['@Neighborhoods\ThrowableDiagnosticComponent\ThrowableDiagnosticV1\Diagnosed\FactoryInterface']]
