+++ src/ThrowableDiagnosticV1Decorators/PostgresV1/PostgresDecorator.buphalo.v1.fabrication.yml
@@ -1,13 +1,13 @@
 actors:
   <PrimaryActorName>.php:
-    template: ThrowableDiagnostic/Decorator/PrimaryActorName.php
+    template: ThrowableDiagnosticComponent/ThrowableDiagnosticV1/DiagnosingDecoratorV1/PrimaryActorName.php
   <PrimaryActorName>.service.yml:
-    template: ThrowableDiagnostic/Decorator/PrimaryActorName.service.yml
+    template: ThrowableDiagnosticComponent/ThrowableDiagnosticV1/DiagnosingDecoratorV1/PrimaryActorName.service.yml
   <PrimaryActorName>Interface.php:
-    template: ThrowableDiagnostic/Decorator/PrimaryActorNameInterface.php
+    template: ThrowableDiagnosticComponent/ThrowableDiagnosticV1/DiagnosingDecoratorV1/PrimaryActorNameInterface.php
   <PrimaryActorName>/AwareTrait.php:
     template: PrimaryActorName/AwareTrait.php
   <PrimaryActorName>/Factory.php:
-    template: ThrowableDiagnostic/Decorator/PrimaryActorName/Factory.php
+    template: ThrowableDiagnosticComponent/ThrowableDiagnosticV1/DiagnosingDecoratorV1/PrimaryActorName/Factory.php
   <PrimaryActorName>/Factory.service.yml:
-    template: ThrowableDiagnostic/Decorator/PrimaryActorName/Factory.service.yml
+    template: ThrowableDiagnosticComponent/ThrowableDiagnosticV1/DiagnosingDecoratorV1/PrimaryActorName/Factory.service.yml
