+++ template-tree/BuphaloV1/ThrowableDiagnosticComponent/ThrowableDiagnosticV1/DecoratorV1/PrimaryActorName.php
@@ -2,22 +2,24 @@
 
 declare(strict_types=1);
 
-namespace Neighborhoods\ThrowableDiagnosticComponent\ThrowableDiagnosticV1Decorators\NestedDiagnosticV1;
+namespace Neighborhoods\BuphaloTemplateTree;
 
-use Neighborhoods\ThrowableDiagnosticComponent\ThrowableDiagnosticV1\DiagnosedInterface;
 use Neighborhoods\ThrowableDiagnosticComponent\ThrowableDiagnosticV1\ThrowableDiagnostic;
 use Neighborhoods\ThrowableDiagnosticComponent\ThrowableDiagnosticV1\ThrowableDiagnosticInterface;
 use Throwable;
 
-final class NestedDiagnosticDecorator implements NestedDiagnosticDecoratorInterface
+final class PrimaryActorName implements PrimaryActorNameInterface
 {
     use ThrowableDiagnostic\AwareTrait;
 
     public function diagnose(Throwable $throwable): ThrowableDiagnosticInterface
     {
-        if ($throwable instanceof DiagnosedInterface) {
-            /** @noinspection PhpUnhandledExceptionInspection */
-            throw $throwable;
+        /** @neighborhoods-buphalo:annotation-processor Neighborhoods\BuphaloTemplateTree\ThrowableDiagnostic\Decorator\PrimaryActorName.diagnose
+        // TODO: Implement diagnose() method.
+        throw new \LogicException('Unimplemented custom diagnostic.');
+         */
+        if (false) {
+            return $this;
         }
 
         $this->getThrowableDiagnosticV1ThrowableDiagnostic()->diagnose($throwable);
