+++ template-tree/BuphaloV1/ThrowableDiagnosticComponent/ThrowableDiagnosticV1/MessageBasedDecoratorV1/PrimaryActorName/Factory.service.yml
@@ -1,0 +1,7 @@
+services:
+  Neighborhoods\BuphaloTemplateTree\PrimaryActorName\FactoryInterface:
+    class: Neighborhoods\BuphaloTemplateTree\PrimaryActorName\Factory
+    public: false
+    shared: true
+    calls:
+      - [setThrowableDiagnosticV1ThrowableDiagnosticDecorator, ['@Neighborhoods\BuphaloTemplateTree\PrimaryActorNameInterface']]
