+++ template-tree/BuphaloV1/ThrowableDiagnosticComponent/ThrowableDiagnosticV1/MessageBasedDiagnosingDecoratorV1/PrimaryActorName.service.yml
@@ -1,0 +1,5 @@
+services:
+  Neighborhoods\BuphaloTemplateTree\PrimaryActorNameInterface:
+    class: Neighborhoods\BuphaloTemplateTree\PrimaryActorName
+    public: false
+    shared: false
