+++ template-tree/BuphaloV1/ThrowableDiagnosticComponent/ThrowableDiagnosticV1/MessageBasedDiagnosingDecoratorV1/PrimaryActorName/Factory.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Neighborhoods\BuphaloTemplateTree\PrimaryActorName;
+
+// @codingStandardsIgnoreLine Line below exceeds 120 characters
+use Neighborhoods\ThrowableDiagnosticComponent\ThrowableDiagnosticV1\ThrowableDiagnostic\Decorator\Factory as DecoratorFactory;
+
+final class Factory extends DecoratorFactory
+{
+}
