+++ template-tree/BuphaloV1/ThrowableDiagnosticComponent/ThrowableDiagnosticV1/MessageBasedDiagnosingDecoratorV1/PrimaryActorNameInterface.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Neighborhoods\BuphaloTemplateTree;
+
+// @codingStandardsIgnoreLine Line below exceeds 120 characters
+use Neighborhoods\ThrowableDiagnosticComponent\ThrowableDiagnosticV1\ThrowableDiagnostic\DecoratorInterface as ThrowableDiagnosticDecoratorInterface;
+
+interface PrimaryActorNameInterface extends ThrowableDiagnosticDecoratorInterface
+{
+
+}
