+++ test/ThrowableDiagnosticV1Decorators/DecoratorTestCase.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Neighborhoods\ThrowableDiagnosticComponentTest\Decorator;
+namespace Neighborhoods\ThrowableDiagnosticComponentTest\ThrowableDiagnosticV1Decorators;
 
 use PHPUnit\Framework\TestCase;
