+++ test/ThrowableDiagnosticV1Decorators/NestedDiagnosticDecoratorTest.php
@@ -1,14 +1,17 @@
 <?php
 
-namespace Neighborhoods\ThrowableDiagnosticComponentTest\Decorator;
+namespace Neighborhoods\ThrowableDiagnosticComponentTest\ThrowableDiagnosticV1Decorators;
 
 use Neighborhoods\ThrowableDiagnosticComponent\ThrowableDiagnosticV1\DiagnosedInterface;
 // @codingStandardsIgnoreLine Line below exceeds 120 characters
 use Neighborhoods\ThrowableDiagnosticComponent\ThrowableDiagnosticV1Decorators\NestedDiagnosticV1\NestedDiagnosticDecorator;
 use Throwable;
+use PHPUnit\Framework\TestCase;
 
-class NestedDiagnosticDecoratorTest extends DecoratorTestCase
+class NestedDiagnosticDecoratorTest extends TestCase
 {
+    use ThrowableDiagnosticMockerTrait;
+
     protected $decorator;
 
     public function setUp(): void
@@ -17,7 +20,6 @@ class NestedDiagnosticDecoratorTest extends DecoratorTestCase
 
         $this->decorator = new NestedDiagnosticDecorator();
         $this->decorator
-            ->setThrowableDiagnosticV1DiagnosedFactory($this->getDiagnosedFactoryMock())
             ->setThrowableDiagnosticV1ThrowableDiagnostic($this->getThrowableDiagnosticMock());
     }
 
@@ -25,7 +27,6 @@ class NestedDiagnosticDecoratorTest extends DecoratorTestCase
     {
         $analysedThrowable = $this->createMock(DiagnosedInterface::class);
         $this->expectNoForwarding();
-        $this->expectNoDiagnosedCreation();
         $this->expectExceptionObject($analysedThrowable);
         $this->decorator->diagnose($analysedThrowable);
     }
@@ -34,7 +35,6 @@ class NestedDiagnosticDecoratorTest extends DecoratorTestCase
     {
         $analysedThrowable = $this->createMock(Throwable::class);
         $this->expectForwarding($analysedThrowable);
-        $this->expectNoDiagnosedCreation();
         $this->decorator->diagnose($analysedThrowable);
     }
 }
