+++ test/ThrowableDiagnosticV1Decorators/SymfonyHttpClientDecoratorTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Neighborhoods\ThrowableDiagnosticComponentTest\Decorator;
+namespace Neighborhoods\ThrowableDiagnosticComponentTest\ThrowableDiagnosticV1Decorators;
 
 // @codingStandardsIgnoreLine Line below exceeds 120 characters
 use Neighborhoods\ThrowableDiagnosticComponent\ThrowableDiagnosticV1Decorators\SymfonyHttpClientV1\SymfonyHttpClientDecorator;
